requirements.txt
selenium==3.141.0
google-cloud-storage==1.38.0
yagmail==0.14.256